bsp/lh2: drv/protocol: carry homographies as float32
The homography travels as IEEE-754 float32 in millimetres instead of an
int32 holding the value times 1e3, so the host packer and every firmware
that links this must move together. An out-of-range basestation index
is now ignored by the store instead of writing past the array.

AI-assisted: Claude Opus 5

+++ bsp/lh2.h
@@ -111,10 +111,10 @@ void db_lh2_calculate_position(uint32_t count1, uint32_t count2, uint32_t basest
  *
  * @param[in] lh2 pointer to the lh2 instance
  * @param[in] basestation_index: index of the basestation that corresponds to the calibration
- * @param[in] homography_matrix_from_packet: the 3x3 homography matrix
+ * @param[in] homography_matrix_from_packet: the 3x3 homography matrix, in millimetres; ignored if basestation_index is out of range
  *
  */
-void db_lh2_store_homography(db_lh2_t *lh2, uint8_t basestation_index, int32_t homography_matrix_from_packet[3][3]);
+void db_lh2_store_homography(db_lh2_t *lh2, uint8_t basestation_index, float homography_matrix_from_packet[3][3]);
 
 /**
  * @brief Handle the fast SPI interrupt
